a fair amount of moving shit around
