+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod error;
 pub mod extend;
+pub mod format;
 pub mod img;
 pub mod ndsfile;
 pub mod proj;
